\name{mederrRank-package}
\alias{mederrRank-package}
\alias{mederrRank}
\docType{package}
\title{
  \packageTitle{mederrRank}
}
\description{
  \packageDescription{mederrRank}
}
\details{
  The package is loaded with the usual \code{library(mederrRank)} command. The most important functions are \code{\link{bhm.mcmc}}, \code{\link{bhm.resample}} and \code{\link{mixnegbinom.em}}.
}
\author{
  \packageAuthor{mederrRank}

  Maintainer: \packageMaintainer{mederrRank}
}
\references{
  Myers, J. A., Venturini, S., Dominici, F. and Morlock, L. (2011), "Random Effects Models for Identifying the Most Harmful Medication Errors in a Large, Voluntary Reporting Database". Technical Report.
}
\seealso{
   \code{\link{bayes.rank}},
   \code{\link{bhm.mcmc}}, 
   \code{\link{bhm.resample}},
   \code{\link{mixnegbinom.em}}.
}
\keyword{Bayesian Hierarchical Models}
\keyword{EM}
\keyword{Empirical Bayes}
\keyword{Mixture Distributions}
\keyword{Optimal Bayes ranking}
